Title: 합과 평균의 차, Time: 140ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240418/합과 평균의 차/sub-of-average-and-sum.py
@@ -1,6 +1,6 @@
 a, b, c = map(int, input().split())
 total = a+b+c
-avg = total/3
+avg = int(total/3)
 
 print(total)
 print(avg)
